fix:fixed the date issues

+++ ACBackendAPI.Domain/Entities/Programme.cs
@@ -3,7 +3,7 @@
     public class Programme : BaseEntity
     {
         public string ProgrammeName { get; set; }
-        public long ProgrammeFee { get; set; }
+        public decimal ProgrammeFee { get; set; }
         public string ProgrammeDescription { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
